delete iklan baris

+++ src/controllers/iklan_baris.js
@@ -2,16 +2,19 @@ const iklanBarisModel = require("../models/iklan_baris");
 const miscHelper = require("../helpers");
 const uniqid = require("uniqid");
 const { ip } = require("../configs");
+const { error } = require("console");
 const filesystem = require("fs").promises;
 
 const deleteFile = async (iklan_baris_id) => {
   const checkId = await iklanBarisModel.checkId(iklan_baris_id);
   const dataIklanBaris = checkId[0];
-  const path = dataIklanBaris.iklan_baris_image.replace(
-    `http://${ip}`,
-    `../backend_lensajabar`
-  );
-  await filesystem.unlink(path);
+  if (dataIklanBaris !== undefined) {
+    const path = dataIklanBaris.iklan_baris_image.replace(
+      `http://${ip}`,
+      `../backend_lensajabar`
+    );
+        await filesystem.unlink(path);
+  }
 };
 module.exports = {
   insertIklanBaris: async (request, response) => {
@@ -132,7 +135,10 @@ module.exports = {
           iklan_baris_sub_district: request.body.iklan_baris_sub_district,
           date_updated: new Date(),
         };
-        const result = await iklanBarisModel.updateIklanBaris(data, iklan_baris_id);
+        const result = await iklanBarisModel.updateIklanBaris(
+          data,
+          iklan_baris_id
+        );
         miscHelper.customResponse(response, 200, result);
       }
     } catch (error) {
@@ -141,6 +147,21 @@ module.exports = {
         response,
         404,
         "Cannot update iklan baris!"
+      );
+    }
+  },
+  deleteIklanBaris: async (request, response) => {
+    try {
+      const iklan_baris_id = request.params.iklan_baris_id;
+      await deleteFile(iklan_baris_id);
+      const result = await iklanBarisModel.deleteIklanBaris(iklan_baris_id);
+      miscHelper.customResponse(response, 200, result);
+    } catch (error) {
+      console.log(error);
+      miscHelper.customErrorResponse(
+        response,
+        404,
+        "Cannot delete iklan baris!"
       );
     }
   },
